UPD: updates the integer interface of the object to use a 64-bit integer because large values may be used when dealing w/ memory attributes (ie.: reading status about JVM memory)

+++ src/base/common/bmic_object.h
@@ -41,7 +41,7 @@ extern "C" {
 
 
 typedef union bmic_value_t_ {
-    int32_t number;
+    int64_t number;
     char *str;
     bool value;
     double d;
